src/worker/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.1.1
dnspython==2.4.1
environs==9.5.0
exceptiongroup==1.1.2
frozenlist==1.4.0
googleapis-common-protos==1.60.0
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.56.0
grpcio-tools==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
loguru==0.7.0
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
pika==1.3.2
pinecone-client==2.2.2
portalocker==2.7.0
protobuf==3.20.2
psycopg2-binary==2.9.6
py==1.11.0
pyarrow==13.0.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pylance==0.8.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qdrant-client==1.4.0
ratelimiter==1.2.0.post0
regex==2023.10.3
requests==2.31.0
retry==0.9.2
semver==3.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tiktoken==0.5.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
validators==0.21.0
weaviate-client==3.22.1
yarl==1.9.2
